Added logic to make reference line in reliability charts come last. Did this by having XML have duplicate XML for the ref line, the first and last XML. So that, when reliability is plotted, the first ref line is drawn with no legend, and last ref line is not drawn, but put in legend. This gets around multiple issues associated with trying to get the ref line last in legend but still plot behind other data lines (especially since a JClass marker can only be a single line at a specific x or y axis value. Closes #1394.
For reliability, now that ref line in legend is last, it makes the columns with separate categories look neater and line up properly and helps a bit with truncating text, although it is still going to truncate if too long (See # 1423). Reliablity diagrams now are set to draw 3 cols instead of 4 (previously due to ref line issues).