Finish working prototype with Plotly plot
The complete lifecycle of making a request to the servlet, getting an XML response, converting the XML to JSON, and sending the data to Plotly, is working.

Next steps are to cleanup the page and (hopefully) optimize some things to make the page load faster (seems a little slow right now).