Produce GeoJSON for the map page
If the user is on the Map page, then produce a GeoJSON. There are different things in this JSON, compared to when the user is on the Chart page. For example, this GeoJSON has `features` array with geospatial data (lat, lon, etc.).

Note that the GeoJSON is produced, but the code to embed the ESRI map, and to plot the data, has not been added yet.