Bugfix - stop legend and averages from cropping
For some reason, when the plot is exported to a PNG, the right side of the legend and, especially, the averages gets cropped. Remove the border around both (which seems to have fixed the issue with the legend) and add a bunch of space to the right side of the averages to the cropping doesn't reach the text.