Bugfix - Fix date axis for long-lead charts
When Monthly or Seasonal forecasts are verified, the resulting scores are all plotted at a single point. It turns out that the `referenceArray` in the XML from the servlet was formatted as `MM/DD/YYYY` for short leads, but `MM/YYYY` for long leads. Just need to take that into account when parsing the dates for charts.