Change how NaNs are removed
In b2ed118 I set scores of NaN to an empty string because for some reason NaNs were being plotted as the value of the previous score in the charts. This line broke the way I *removed* NaNs completely in `plotting.js`. This change now removes stations/grid points where the score is empty.