Fix reliability line colors
When the black perfect reliability line is inserted at the beginning of the Plotly data object (so that it's plotted first and doesn't hide the other lines), it shifts the colors of the proceeding lines.

I just created an array of 4 colors and manually set the colors of each forecast skill line to that color.

Fixes #24