Fix a bug where weekend data is skipped for manual forecasts
The `mean()` function calculates the mean of an array, but to skip missing data it actually removes the data from the array, shortening it. This was being done on the array being passed in (by reference), meaning the original array was being modified. I instead made a shallow copy of the input array in the `mean()` function first, then did all work on that array instead.

Fixes #26