Fix date comparison subroutine
The subroutine `compareDates()` is used to check the start and end dates
provided and make sure the start date comes before the end date.
Depending on whether the dates were 6 or 8 characters long, the
appropriate date format was being calculated, but nothing was ever done
with that, so regardless of whether they were 6 or 8 digit dates they
were being passed to `ParseDate()` as is, which works for 8-digit dates
but not 6-digit dates.