Fix perfect reliability line
For monthly and seasonal reliability in the VWT, there was a 0.3333 in
the x-axis, but the Y values supplied to Plot.ly to create the perfect
reliability line was missing that 0.3333, so the perfect reliability
line wasn't correct after it passed 0.3333. I now remove 0.3333 from the
arrays received by the VWT, which we really don't want to plot anyway.